--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -10,13 +10,7 @@ import (
 
 //easyjson:json
 type User struct {
-	//ID       int
-	//Name     string
-	//Username string
 	Email string
-	//Phone    string
-	//Password string
-	//Address  string
 }
 
 type DomainStat map[string]int